+++ src/main/java/racingcar/model/race/CarRaceRoundManager.java
@@ -1,4 +1,4 @@
-package racingcar.model;
+package racingcar.model.race;
 
 public class CarRaceRoundManager {
