+++ src/main/java/racingcar/model/race/strategy/MoveStrategy.java
@@ -1,4 +1,4 @@
-package racingcar.model.move;
+package racingcar.model.race.strategy;
 
 public interface MoveStrategy {
